+++ src/config/defaultSetting.ts
@@ -6,6 +6,7 @@ const defaultSetting: LX.AppSetting = {
   'common.sourceNameType': 'alias',
   'common.shareType': 'system',
   'common.isAgreePact': false,
+  'common.isUserApiDefaultsImported': false,
   'common.autoHidePlayBar': true,
   'common.drawerLayoutPosition': 'left',
   'common.homePageScroll': true,