+++ src/screens/Home/Views/Setting/settings/Basic/UserApiEditModal/index.tsx
@@ -9,6 +9,7 @@ import Dialog, { type DialogType } from '@/components/common/Dialog'
 import Button from '@/components/common/Button'
 import List from './List'
 import ImportBtn from './ImportBtn'
+import ScriptImportDefault from './ScriptImportDefault'
 
 // interface UrlInputType {
 //   setText: (text: string) => void
@@ -125,6 +126,7 @@ export default forwardRef<UserApiEditModalType, {}>((props, ref) => {
                 <Text size={14} color={theme['c-button-font']}>{t('close')}</Text>
               </Button>
               <ImportBtn btnStyle={{ ...styles.btn, backgroundColor: theme['c-button-background'] }} />
+              <ScriptImportDefault />
             </View>
           </Dialog>
         ) : null